Convert everything but commands to TS
