+++ package.json
@@ -40,11 +40,12 @@
 		"mariadb": "^3.0.0",
 		"mongoose": "^6.0.12",
 		"mqtt": "^4.2.8",
-		"node-persist": "^3.1.0",
+		"node-persist": "^3.1.3",
 		"ts-node": "^10.9.1",
 		"typescript": "^5.0.4"
 	},
 	"devDependencies": {
+		"@types/node-persist": "^3.1.3",
 		"eslint": "^8.23.1",
 		"eslint-config-airbnb-base": "^15.0.0",
 		"eslint-plugin-import": "^2.25.2",
