+++ src/printer_events/PrintPaused.ts
@@ -1,27 +1,27 @@
-const printers = require('../lib/printers');
-const ldap = require('../lib/ldap');
-const Discord = require('discord.js');
+import printers = require('../lib/printers');
+import ldap = require('../lib/ldap');
+import * as Discord from 'discord.js';
 
 module.exports = {
-	name: 'PrintCancelled',
-	execute: async (data, printerId, client) => {
-		const ldapUser = await ldap.getUserByUsername(data.owner, 'discord');
+	name: 'PrintPaused',
+	execute: async (data: any, printerId: string, client: Discord.Client) => {
+		const ldapUser = await ldap.getUserByUsername(data.owner, ['discord']);
 
 		// If there is no user, we can't do anything. Otherwise, get the user.
 		if (ldapUser && ldapUser.discord) {
-			const user = await client.users.fetch(ldapUser.discord);
+			const user = await client.users.fetch(ldapUser.discord as string);
 			if (user) {
 				// Construct our embed.
 				const embed = await printers.getEmbedTemplate(printerId);
 				embed
-					.setTitle('⚠  Print Cancelled')
+					.setTitle('🍝  Print Paused')
 					.setDescription(
-						`Your print was cancelled <t:${Math.round(
+						`Your print was paused <t:${Math.round(
 							Date.now() / 1000
-						)}:R>. This can happen for a variety of reasons, such as spaghetti being detected or your print violating our rules. Please come to the UTDesign Makerspace to clean up the printer as soon as possible.`
+						)}:R>. This usually happens because a spaghetti printing fail was detected. Please come to the UTDesign Makerspace to either resume or cancel your print as soon as possible.`
 					)
 					.setTimestamp()
-					.setColor('#dd2e44');
+					.setColor('#fdcb58');
 
 				// Add image to embed.
 				const snapshotBuffer = await printers.getSnapshotBuffer(
@@ -37,7 +37,10 @@ module.exports = {
 
 				// Send the embed to the user.
 				await user
-					.send({ embeds: [embed], files: [snapshot] })
+					.send({
+						embeds: [embed],
+						files: snapshot ? [snapshot] : []
+					})
 					.catch(() => {
 						// If the user is not accepting DMs, we can't send them the message.
 						console.log(
@@ -50,3 +53,5 @@ module.exports = {
 		printers.updateWatcher(printerId, client);
 	}
 };
+
+export {};
